Revert "server clean up"
